+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 ### Added
 
 - Add utf8 encoding while loading datamodel
+- GeoDjango PointField to compute distances
 
 ### Changed